worker/requirements.txt
rq==2.3.2
python-dotenv==1.1.0